LayoutTests/editing/editability/attr-missing-parent-true.html

<html>
<head>
<link rel="stylesheet" href="../../fast/js/resources/js-test-style.css">
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></p>
<div id="result"></div>
<div id="div" contentEditable="true">
    <p id="p"></p>
</div>
<script>
description('When contentEditable is missing, it maps to "inherit" state and the element is editable if its parent is.')

shouldBe('document.getElementById("p").hasAttribute("contentEditable")','false');
shouldBe('document.getElementById("p").contentEditable', '"inherit"');
debug("FIXME: contentEditable attribute should be inherit if missing. https://bugs.webkit.org/show_bug.cgi?id=52056");
shouldBe('document.getElementById("p").isContentEditable', 'true');
shouldBe('window.getComputedStyle(p, "").getPropertyValue("-webkit-user-modify")', '"read-write"');

document.getElementById("div").style.display= 'none';

window.successfullyParsed = true;
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
